+++ examples/expo-swr/.wundergraph/wundergraph.config.ts
@@ -4,7 +4,7 @@ import operations from './wundergraph.operations';
 
 const spaceX = introspect.graphql({
 	apiNamespace: 'spacex',
-	url: 'https://api.spacex.land/graphql/',
+	url: 'https://spacex-api.fly.dev/graphql/',
 });
 
 // configureWunderGraph emits the configuration
@@ -35,9 +35,6 @@ configureWunderGraphApplication({
 						'http://localhost:3000',
 				  ]
 				: ['http://localhost:3000'],
-	},
-	dotGraphQLConfig: {
-		hasDotWunderGraphDirectory: false,
 	},
 	security: {
 		enableGraphQLEndpoint: process.env.NODE_ENV !== 'production',