+++ examples/graphql-ws-subscriptions/.wundergraph/wundergraph.config.ts
@@ -40,9 +40,6 @@ configureWunderGraphApplication({
 		...cors.allowAll,
 		allowedOrigins: [],
 	},
-	dotGraphQLConfig: {
-		hasDotWunderGraphDirectory: false,
-	},
 	security: {
 		enableGraphQLEndpoint: process.env.NODE_ENV !== 'production' || process.env.GITPOD_WORKSPACE_ID !== undefined,
 	},