+++ packages/sdk/src/client/client.ts
@@ -21,6 +21,13 @@ import { InputValidationError } from './InputValidationError';
 
 // https://graphql.org/learn/serving-over-http/
 
+export interface UploadValidationOptions {
+	maxAllowedUploadSizeBytes?: number;
+	maxAllowedFiles?: number;
+	allowedFileExtensions?: string[];
+	allowedMimeTypes?: string[];
+}
+
 interface LogoutResponse {
 	redirect?: string;
 }
@@ -230,12 +237,7 @@ export class Client {
 
 		const headers: Headers = {};
 
-		if (
-			this.options.operationMetadata &&
-			this.options.operationMetadata[options.operationName] &&
-			this.options.operationMetadata[options.operationName].requiresAuthentication &&
-			this.csrfEnabled
-		) {
+		if (this.isAuthenticatedOperation(options.operationName) && this.csrfEnabled) {
 			headers['X-CSRF-Token'] = await this.getCSRFToken();
 		}
 
@@ -366,7 +368,11 @@ export class Client {
 	 * could not be uploaded for any reason. If the upload was successful, your return a list
 	 * of file IDs that can be used to download the files from your S3 bucket.
 	 */
-	public async uploadFiles(config: UploadRequestOptions): Promise<UploadResponse> {
+	public async uploadFiles<UploadOptions extends UploadRequestOptions>(
+		config: UploadOptions,
+		validation?: UploadValidationOptions
+	): Promise<UploadResponse> {
+		this.validateFiles(config, validation);
 		const formData = new FormData();
 		for (const [_, file] of Object.entries(config.files)) {
 			if (file instanceof Blob) {
@@ -383,6 +389,14 @@ export class Client {
 		const params = new URLSearchParams({
 			wg_api_hash: this.options.applicationHash,
 		});
+
+		if ('profile' in config) {
+			headers['X-Upload-Profile'] = (config as any).profile;
+		}
+
+		if ('meta' in config) {
+			headers['X-Metadata'] = (config as any).meta ? JSON.stringify((config as any).meta) : '';
+		}
 
 		const response = await this.fetch(
 			this.addUrlParams(`${this.options.baseURL}/s3/${config.provider}/upload`, params),
@@ -409,6 +423,42 @@ export class Client {
 		return {
 			fileKeys: json.map((x) => x.key),
 		};
+	}
+
+	public validateFiles(config: UploadRequestOptions, validation?: UploadValidationOptions) {
+		if (validation?.maxAllowedFiles && config.files.length > validation.maxAllowedFiles) {
+			throw new Error(`uploading ${config.files.length} exceeds the maximum allowed (${validation.maxAllowedFiles})`);
+		}
+		for (const file of config.files) {
+			if (validation?.maxAllowedUploadSizeBytes && file.size > validation.maxAllowedUploadSizeBytes) {
+				throw new Error(
+					`file ${file.name} with size ${file.size} exceeds the maximum allowed (${validation.maxAllowedUploadSizeBytes})`
+				);
+			}
+			if (validation?.allowedFileExtensions && file.name.includes('.')) {
+				const ext = file.name.substring(file.name.indexOf('.') + 1).toLowerCase();
+				if (ext) {
+					if (validation.allowedFileExtensions.findIndex((item) => item.toLocaleLowerCase()) < 0) {
+						throw new Error(`file ${file.name} with extension ${ext} is not allowed`);
+					}
+				}
+			}
+			if (validation?.allowedMimeTypes) {
+				const mimeType = file.type;
+				const idx = validation.allowedMimeTypes.findIndex((item) => {
+					// Full match
+					if (item == mimeType) {
+						return true;
+					}
+					// Try wildcard match. This is a bit brittle but it should be fine
+					// as long as profile?.allowedMimeTypes contains only valid entries
+					return mimeType.match(new RegExp(item.replace('*', '.*')));
+				});
+				if (idx < 0) {
+					throw new Error(`file ${file.name} with MIME type ${mimeType} is not allowed`);
+				}
+			}
+		}
 	}
 
 	public login(authProviderID: string, redirectURI?: string) {