+++ packages/sdk/src/client/types.ts
@@ -21,12 +21,21 @@ export interface ClientOperation {
 	requiresAuthentication: boolean;
 }
 
+export interface S3ProviderDefinition {
+	[provider: string]: {
+		hasProfiles: boolean;
+		profiles: {
+			[profile: string]: object;
+		};
+	};
+}
+
 export interface OperationsDefinition<
 	Queries extends OperationDefinition = OperationDefinition,
 	Mutations extends OperationDefinition = OperationDefinition,
 	Subscriptions extends OperationDefinition = OperationDefinition,
 	UserRole extends string = string,
-	S3Provider extends string = string,
+	S3Provider extends S3ProviderDefinition = S3ProviderDefinition,
 	AuthProvider extends string = string
 > {
 	user: User<UserRole>;
@@ -118,6 +127,18 @@ export interface UploadRequestOptions<ProviderName extends string = string> {
 	provider: ProviderName;
 	files: FileList;
 	abortSignal?: AbortSignal;
+}
+
+export interface UploadRequestOptionsWithProfile<
+	ProviderName extends string = string,
+	ProfileName extends string = string,
+	Meta extends object = object
+> extends UploadRequestOptions<ProviderName> {
+	provider: ProviderName;
+	profile: ProfileName;
+	files: FileList;
+	abortSignal?: AbortSignal;
+	meta?: Meta;
 }
 
 export interface UploadResponse {