+++ packages/sdk/src/server/plugins/hooks.ts
@@ -5,6 +5,8 @@ import {
 	HooksConfigurationOperationType,
 	OperationHookFunction,
 	OperationHooksConfiguration,
+	UploadHooks,
+	WunderGraphFile,
 	WunderGraphRequest,
 	WunderGraphResponse,
 } from '../types';
@@ -471,6 +473,73 @@ const FastifyHooksPlugin: FastifyPluginAsync<FastifyHooksOptions> = async (fasti
 		});
 	}
 
+	const preUpload = (providerName: string, profileName: string, handler: any) => {
+		fastify.post<{
+			Body: {
+				file: WunderGraphFile;
+				meta: any;
+			};
+		}>(`/upload/${providerName}/${profileName}/preUpload`, async (request, reply) => {
+			reply.type('application/json').code(200);
+			try {
+				const result = await handler({
+					...request.ctx,
+					file: request.body.file,
+					meta: request.body.meta,
+				});
+				return result || {};
+			} catch (err) {
+				request.log.error(err);
+				reply.code(500);
+				return { error: err };
+			}
+		});
+	};
+
+	const postUpload = (providerName: string, profileName: string, handler: any) => {
+		fastify.post<{
+			Body: {
+				file: WunderGraphFile;
+				meta: any;
+				error: Error;
+			};
+		}>(`/upload/${providerName}/${profileName}/postUpload`, async (request, reply) => {
+			reply.type('application/json').code(200);
+			try {
+				const result = await handler({
+					...request.ctx,
+					file: request.body.file,
+					meta: request.body.meta,
+					error: request.body.error,
+				});
+				return result || {};
+			} catch (err) {
+				request.log.error(err);
+				reply.code(500);
+				return { error: err };
+			}
+		});
+	};
+
+	function registerUploadHooks(hooks: UploadHooks): number {
+		let count = 0;
+		for (const providerName in hooks) {
+			const provider = hooks[providerName];
+			for (const profileName in provider) {
+				const profile = provider[profileName];
+				if (profile?.preUpload !== undefined) {
+					count++;
+					preUpload(providerName, profileName, profile.preUpload);
+				}
+				if (profile?.postUpload !== undefined) {
+					count++;
+					postUpload(providerName, profileName, profile.postUpload);
+				}
+			}
+		}
+		return count;
+	}
+
 	// queries
 	const queryOperations = config?.[HooksConfigurationOperationType.Queries];
 	if (queryOperations) {
@@ -490,6 +559,13 @@ const FastifyHooksPlugin: FastifyPluginAsync<FastifyHooksOptions> = async (fasti
 	if (subscriptionOperations) {
 		registerOperationHooks(subscriptions, subscriptionOperations);
 		fastify.log.debug(`Registered (${subscriptions.length}) subscription operations`);
+	}
+
+	// uploads
+	const uploadOperations = config?.[HooksConfigurationOperationType.Uploads];
+	if (uploadOperations) {
+		const registered = registerUploadHooks(uploadOperations);
+		fastify.log.debug(`Registered (${registered}) upload hooks`);
 	}
 };